requirements.txt
aiofiles==24.1.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
beautifulsoup4==4.12.2
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
Django==5.2
django-scheduler==0.10.1
fastapi==0.110.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
icalendar==6.1.3
idna==3.10
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
playwright==1.52.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
rnet==2.2.8
schedule==1.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.36.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.27.1
